pictures added and doctor component updated
