+++ src/app/Modules/Doctor/doctor landing page/doctor.component.css
@@ -1,6 +1,6 @@
 #docsection {
     height: 750px;
-    background: url("src/assets/img/doc1.png");
+    background: url("src/assets/img/doc3.png");
     /* background-size: 100%; */
     background-size: cover;
     background-position: center;
