+++ src/app/app.module.ts
@@ -17,7 +17,7 @@ import { DoctorComponent } from './Modules/Doctor/doctor landing page/doctor.com
 import { DoctorSignupComponent } from './Modules/Doctor/doctor-signup-page-1/doctor-signup.component';
 import { DoctorSignupNextComponent } from './Modules/Doctor/doctor-signup-page-2/doctor-signup-next.component';
 import { LoginComponent } from './Shared/login/login.component';
-import { HomeComponent } from './home/home.component';
+import { HomeComponent } from './Shared/home/home.component';
 
 @NgModule({
   declarations: [
